Move V2 Project dashboard state to hook
ahhh im abstractinggg
